requirements.txt
selenium==3.141.0
pytest==6.2.2
pytest-xdist==2.2.1
jproperties==2.1.0
allure-pytest==2.8.36